Rename endpoint_peering to remote_ops
