+++ psbench/benchmarks/remote_ops/__main__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from psbench.benchmarks.endpoint_peering.main import main
+from psbench.benchmarks.remote_ops.main import main
 
 if __name__ == '__main__':
     raise SystemExit(main())
